you can add your own beatmap to the dircetory named "songs"

+++ note.h
@@ -7,7 +7,7 @@ class note
 public:
     note();
     int key;
-    int start_time;
+   long int start_time;
     double speed;
 };
